Remove all references to __FILE__ in ncgen
re: issue https://github.com/Unidata/netcdf-c/issues/1323
    and PR https://github.com/Unidata/netcdf-c/pull/1326

Replace the uses of the __FILE__ macro in ncgen with the
constant string "ncgen" in order to keep external package
builders (like Debian) happy.

Note that __FILE__ will still appear in ncgen generated C code
files, but this should be ok since it is not part of the distro
per-se.