Fix includes/netcdf_xxx.h files to not include ncexternl.h
re: https://github.com/Unidata/netcdf-c/issues/1502

If you plan to use any of the netcdf_XXX.h files, then it it is assumed
that you first include netcdf.h, followed by any netcdf_XXX.h files
you need. The EXTERNL (and other things) are defined in netcdf.h
so they should be available for use by the netcdf_XXX.h files.

This PR ensures that the above assumption is true.  It also adds
some comments to that effect to the start of the netcdf_xxx.h files.