Fix reclamation of the ->format_XXX_info fields
Currently, the code assumes that all the values assigned
to the format_XXX_info fields (e.g. format_file_info)
are assigned by libhdf5 or libhdf4. But now nczarr is using that field
and requires nczarr specific free'ing of the format_XXX_info fields.

This solution is to pass a function that can free any of the
format_XXX_info fields. This is primarily passed to the
nc4_XXX_list_del functions.