Make fillmismatch the default for DAP2 and DAP4
re: https://github.com/Unidata/netcdf-c/issues/1614

NetCDF has a requirement that the type of a _FillValue attribute
be the same as the containing variable.  However, it is clear
that too many servers do not honor this requirement.  So, change
the default for DAP2 and DAP4 to allow fill mismatch.