Fix if statement to apply to fflush
Even though the `fflush()` is on the same line as the `fprintf(stderr, ...` statement, it is not part of the `if` and is therefore, executed even if the `wdebug` is not active.  This results in `fflush()` being called more than it should.  

Added parenthesis to property protect the fflush call.