Fix some HDF5-related tests
Some of the tests for HDF5 symbols are failing due to not finding the include file which skews the test results.  Added the `CMAKE_REQUIRED_INCLUDES` command to set the path.  In one place this was already set, but used `HAVE_HDF5_H` which is not an include path and isn't even set at the location where it was used.  We may not need these set more than once, but I added them before each `CHECK_C_SOURCE_COMPILES` in case those blocks get moved around and the value gets unset when it should be set.

Also changed the `H5Literate_vers` check to check  that it is defined instead of checking the value.  As best I can see, as long as it is defined, then the `H5Literate` symbol will be defined which is what is needed.  Checked with 1.10.7 and 1.12.0